+++ aries-site/src/tests/accessibility/README.md
@@ -1,0 +1,7 @@
+# Accessibility Testing
+
+### What does this folder test?
+
+This folder tests accessibility features such as aria features, color contrast, and any other rules supported by [axe](https://dequeuniversity.com/rules/axe/3.5). For keyboard accessibility tests, see the files in `tests/navigation/`.
+
+We are using [axe-testcafe](https://www.npmjs.com/package/axe-testcafe) to integrate these checks with TestCafe.
